#32 Made the 12 Days a generic page
The 12 Days can now be passed a more detailed JSON file to render a new version. This will allow future 12 Day content to be pushed much easier.
