+++ components/12-days/Flap.js
@@ -1,17 +1,13 @@
 import React from 'react';
 import IcomoonReact from "icomoon-react";
 import iconSet from "../icons/selection.json";
-import Data_EN from './Data';
-import Data_FR from './Data-fr';
 import styles from './Flap.module.css';
 
 function Flap(props) {
 
-    // let url = window.location !== window.parent.location ? document.referrer : document.location.href;
-    // let Data = url.includes("/fr/") ? Data_FR : Data_EN;
-    let Data = Data_EN;
+    console.log(props);
 
-    const dayContent = Data[`day_${props.day}`];
+    const dayContent = props.data.content[`day_${props.day}`];
 
     function open(ev) {
         if (checkDate()){
